+++ src/util/interpolate.ts
@@ -3,7 +3,8 @@ import Color from './color';
 import Padding from './padding';
 import VariableAnchorOffsetCollection from './variable_anchor_offset_collection';
 import RuntimeError from '../expression/runtime_error';
-import {VariableAnchorOffsetCollectionSpecification} from '../types.g';
+import type {VariableAnchorOffsetCollectionSpecification} from '../types.g';
+import Projection from './projection';
 
 export type InterpolationColorSpace = 'rgb' | 'hcl' | 'lab';
 
@@ -23,10 +24,11 @@ export function isSupportedInterpolationColorSpace(colorSpace: string): colorSpa
  * @returns interpolation fn
  * @deprecated use `interpolate[type]` instead
  */
-export const interpolateFactory = (interpolationType: 'number'|'color'|'array'|'padding'|'variableAnchorOffsetCollection') => {
+export const interpolateFactory = (interpolationType: 'number'|'color'|'projection'|'array'|'padding'|'variableAnchorOffsetCollection') => {
     switch (interpolationType) {
         case 'number': return number;
         case 'color': return color;
+        case 'projection': return projection;
         case 'array': return array;
         case 'padding': return padding;
         case 'variableAnchorOffsetCollection': return variableAnchorOffsetCollection;
@@ -35,6 +37,10 @@ export const interpolateFactory = (interpolationType: 'number'|'color'|'array'|'
 
 function number(from: number, to: number, t: number): number {
     return from + t * (to - from);
+}
+
+function projection(from: string, to: string, interpolation: number): Projection {
+    return new Projection(from, to, interpolation);
 }
 
 function color(from: Color, to: Color, t: number, spaceKey: InterpolationColorSpace = 'rgb'): Color {
@@ -122,6 +128,7 @@ function variableAnchorOffsetCollection(from: VariableAnchorOffsetCollection, to
 const interpolate = {
     number,
     color,
+    projection,
     array,
     padding,
     variableAnchorOffsetCollection